+++ src/exports.zig
@@ -4,6 +4,7 @@ pub const core = struct {
     pub const debug = @import("core/debug.zig");
     pub const buffer_controller = @import("core/buffer_controller.zig");
     pub const data_paths = @import("core/data_paths.zig");
+    pub const data_location = @import("core/data_location.zig");
     pub const runtime_settings = @import("core/runtime_settings.zig");
     pub const application_exclusions = @import("core/application_exclusions.zig");
 };